+++ pattern/test/sg/flybot/integrated_test.clj
@@ -1,7 +1,0 @@
-(ns sg.flybot.integrated-test
-  (:require
-   [clojure.test :refer [deftest]]
-   [com.mjdowney.rich-comment-tests.test-runner :as tr]))
-
-(deftest rct-tests
-  (tr/run-tests-in-file-tree! :dirs #{"src"}))